unirecast stopping point

+++ AINavigationML/UniRecast.Core/UniRcNavMeshSurface.cs
@@ -5,6 +5,7 @@ using DotRecast.Detour;
 using DotRecast.Recast;
 using DotRecast.Recast.Toolset;
 using Il2CppGB.Game;
+using Il2CppGB.Game.PlayArea;
 using MelonLoader;
 using MelonLoader.Utils;
 using UnityEngine;
@@ -24,7 +25,7 @@ namespace UniRecast.Core
         // Agent
         private float _agentHeight = 1.4f;
 
-        private float _agentRadius = 0.1f;
+        private float _agentRadius = 0.4f;
 
         private float _agentMaxClimb = 0.9f;
 
@@ -47,11 +48,11 @@ namespace UniRecast.Core
         private bool _filterWalkableLowHeightSpans = true;
 
         // Polygonization
-        private float _edgeMaxLen = 6f;
+        private float _edgeMaxLen = 12f;
 
-        private float _edgeMaxError = 0.5f;
+        private float _edgeMaxError = 1f;
 
-        private int _vertsPerPoly = 3;
+        private int _vertsPerPoly = 6;
 
         // Detail Mesh
         private float _detailSampleDist = 6f;
@@ -59,7 +60,7 @@ namespace UniRecast.Core
         private float _detailSampleMaxError = 1f;
 
         // Tiles
-        private int _tileSize = 64;
+        private int _tileSize = 32;
 
         private UniRcNavMeshData _navMeshData = new();
 
@@ -85,9 +86,9 @@ namespace UniRecast.Core
             bs.mergedRegionSize = _mergedRegionSize;
 
             // Filtering
-            bs.filterLowHangingObstacles = false;
-            bs.filterLedgeSpans = false;
-            bs.filterWalkableLowHeightSpans = false;
+            bs.filterLowHangingObstacles = _filterLowHangingObstacles;
+            bs.filterLedgeSpans = _filterLedgeSpans;
+            bs.filterWalkableLowHeightSpans = _filterWalkableLowHeightSpans;
 
             // Polygonization
             bs.edgeMaxLen = _edgeMaxLen;
@@ -106,6 +107,11 @@ namespace UniRecast.Core
             bs.partitioning = RcPartitionType.MONOTONE.Value;
 
             return bs;
+        }
+
+        private void Update()
+        {
+            _navMeshData.dynamicNavMesh?.Update();
         }
 
         public void Bake(bool useScene = true)
@@ -129,44 +135,52 @@ namespace UniRecast.Core
             var currentScene = SceneManager.GetActiveScene();
             var navMeshFilePath = Path.Combine(MelonEnvironment.UserDataDirectory, currentScene.name + ".bytes");
 
-            if (File.Exists(navMeshFilePath))
+            if (useScene)
             {
-                try
+                if (File.Exists(navMeshFilePath))
                 {
-                    _navMeshData.NavMesh = UniRcExtensions.LoadNavMeshFile(navMeshFilePath);
+                    try
+                    {
+                        _navMeshData.dynamicNavMesh.NavMesh(UniRcExtensions.LoadNavMeshFile(navMeshFilePath));
+                        return;
+                    }
+                    catch (Exception e)
+                    {
+                        Entrypoint.Logger.Error(e);
+                        File.Delete(navMeshFilePath);
+                    }
+                }
+
+                GameObject[] allGameObjects = currentScene.GetRootGameObjects();
+
+                var targets = GetNavMeshSurfaceTargets(allGameObjects);
+                if (0 >= targets.Count)
+                {
+                    Entrypoint.Logger.Msg($"not found navmesh targets");
                     return;
                 }
-                catch (Exception e)
-                {
-                    Entrypoint.Logger.Error(e);
-                    File.Delete(navMeshFilePath);
-                }
-            }
 
-            GameObject[] allGameObjects = currentScene.GetRootGameObjects();
-
-            var targets = GetNavMeshSurfaceTargets(allGameObjects);
-            if (0 >= targets.Count)
-            {
-                Entrypoint.Logger.Msg($"not found navmesh targets");
-                return;
-            }
-
-            var combinedTarget = targets.ToCombinedNavMeshSurfaceTarget(currentScene.name);
-            var mesh = combinedTarget.ToMesh();
-            mesh.SaveFile();
+                var combinedTarget = targets.ToCombinedNavMeshSurfaceTarget(currentScene.name);
+                var mesh = combinedTarget.ToMesh();
+                mesh.SaveFile();
             
-            var debugDisplayObj = new GameObject();
-            var debugDisplayFilter = debugDisplayObj.AddComponent<MeshFilter>();
-            debugDisplayObj.AddComponent<MeshRenderer>();
-            debugDisplayFilter.sharedMesh = combinedTarget.GetMesh();
+                var debugDisplayObj = new GameObject();
+                var debugDisplayFilter = debugDisplayObj.AddComponent<MeshFilter>();
+                debugDisplayObj.AddComponent<MeshRenderer>();
+                debugDisplayFilter.sharedMesh = combinedTarget.GetMesh();
 
-            var navMesh = mesh.Build(setting);
-            _navMeshData.NavMesh = navMesh;
+                var navMesh = mesh.Build(setting);
+                _navMeshData.dynamicNavMesh = navMesh;
 
-            navMesh.SaveNavMeshFile(navMeshFilePath);
+                (_navMeshData.NavMesh ?? throw new NullReferenceException("NavMesh is null!")).SaveNavMeshFile(navMeshFilePath);
 
-            Entrypoint.Logger.Msg($"{null != _navMeshData}");
+                Entrypoint.Logger.Msg($"New NavMesh baked. | Path: {navMeshFilePath} | Max Tiles: {navMesh.NavMesh().GetMaxTiles()}");
+            }
+            else
+            {
+                var toTarget = gameObject;
+                
+            }
         }
 
         public List<UniRcNavMeshSurfaceTarget> GetNavMeshSurfaceTargets(IList<GameObject> gameObjects)
@@ -178,36 +192,35 @@ namespace UniRecast.Core
                 .Where(x => x.gameObject.activeSelf && x.isActiveAndEnabled)
                 .Select(x => x.ToUniRcSurfaceSource());
                 */
-
+            
             // all tag objects
             var navmeshTagObjects = gameObjects
                 .SelectMany(x => x.ToHierarchyList())
-                .Where(x => x.gameObject.activeSelf)
+                .Where(x => x.gameObject.activeSelf && PlayAreaCache.InPlayArea(x.transform.position) != null)
                 .ToList();
-/*
+            
             var meshFilterTargets = navmeshTagObjects
-                .SelectMany(x => x.GetComponentsInChildren<MeshFilter>().Where(m => m.GetComponent<MeshRenderer>().enabled))
+                .SelectMany(x => x.GetComponentsInChildren<MeshFilter>().Where(m => m.GetComponent<MeshRenderer>().enabled && m.GetComponentsInChildren<Collider>().Any(c => c.enabled && !c.isTrigger)))
                 .Distinct()
                 .Select(x => x.ToUniRcSurfaceSource());
-                */
 
-            var boxColliderTargets = (from obj in navmeshTagObjects
+            /*var boxColliderTargets = (from obj in navmeshTagObjects
                 select obj.GetComponentsInChildren<BoxCollider>()
                 into colliders
                 from eligibleCollider in colliders.Where<BoxCollider>(c => !c.CompareTag("Helper (Kill Volume)") && c.enabled && !c.isTrigger)
-                select eligibleCollider.ToUniRcSurfaceSource()).ToList();
+                select eligibleCollider.ToUniRcSurfaceSource()).ToList();*/
 
             var targets = new List<UniRcNavMeshSurfaceTarget>();
             //targets.AddRange(terrainTargets);
-            //targets.AddRange(meshFilterTargets);,
-            targets.AddRange(boxColliderTargets);
+            targets.AddRange(meshFilterTargets);
+            //targets.AddRange(boxColliderTargets);
 
             return targets;
         }
 
         public void Clear()
         {
-            _navMeshData.NavMesh = null;
+            _navMeshData.dynamicNavMesh = null;
         }
 
         public bool HasNavMeshData()
@@ -215,7 +228,7 @@ namespace UniRecast.Core
             return null != _navMeshData.NavMesh;
         }
 
-        public DtNavMesh GetNavMeshData()
+        public DtNavMesh? GetNavMeshData()
         {
             return _navMeshData.NavMesh;
         }
